Added basic unit test for pulsar buffer and made tweaks to the code itself.

+++ lib/utils/pulsarTiming.hpp
@@ -10,7 +10,7 @@ class Polyco {
 public:
     Polyco(double tmid, float dm, double phase_ref, double rot_freq, std::vector<float> coeff);
 
-    double mjd2phase(float t);
+    double mjd2phase(double t);
 
     double unix2phase(timespec t);
